+++ packages/core/src/storage/vectorStore/ChromaVectorStore.ts
@@ -106,12 +106,32 @@ export class ChromaVectorStore implements VectorStore {
       throw new Error("ChromaDB does not support querying by mode");
     }
 
-    const chromaWhere: { [x: string]: string | number | boolean } = {};
+    // fixme: type is broken
+    let chromaWhere: any = {};
     if (query.filters?.filters) {
       query.filters.filters.map((filter) => {
         const filterKey = filter.key;
         const filterValue = filter.value;
-        chromaWhere[filterKey] = filterValue;
+        if (filterKey in chromaWhere || "$or" in chromaWhere) {
+          if (!chromaWhere["$or"]) {
+            chromaWhere = {
+              $or: [
+                {
+                  ...chromaWhere,
+                },
+                {
+                  [filterKey]: filterValue,
+                },
+              ],
+            };
+          } else {
+            chromaWhere["$or"].push({
+              [filterKey]: filterValue,
+            });
+          }
+        } else {
+          chromaWhere[filterKey] = filterValue;
+        }
       });
     }
 
@@ -130,6 +150,7 @@ export class ChromaVectorStore implements VectorStore {
         IncludeEnum.Embeddings,
       ],
     });
+
     const vectorStoreQueryResult: VectorStoreQueryResult = {
       nodes: queryResponse.ids[0].map((id, index) => {
         const text = (queryResponse.documents as string[][])[0][index];